chore: add test for counterfactual address
