+++ tests/modules/write.test.ts
@@ -10,8 +10,7 @@ import {
   parseEther,
   parseUnits,
   slice,
-  toFunctionSelector,
-  getContract
+  toFunctionSelector
 } from "viem"
 import { privateKeyToAccount } from "viem/accounts"
 import { beforeAll, describe, expect, test } from "vitest"
@@ -19,8 +18,7 @@ import {
   type BiconomySmartAccountV2,
   type Transaction,
   type WalletClientSigner,
-  createSmartAccountClient,
-  DEFAULT_BICONOMY_FACTORY_ADDRESS
+  createSmartAccountClient
 } from "../../src/account"
 import { Logger, getChain } from "../../src/account"
 import {
@@ -28,7 +26,6 @@ import {
   DEFAULT_MULTICHAIN_MODULE,
   DEFAULT_SESSION_KEY_MANAGER_MODULE,
   createBatchedSessionRouterModule,
-  createECDSAOwnershipValidationModule,
   createMultiChainValidationModule,
   createSessionKeyManagerModule,
   getABISVMSessionKeyData
@@ -36,11 +33,8 @@ import {
 import { SessionMemoryStorage } from "../../src/modules/session-storage/SessionMemoryStorage"
 import { PaymasterMode } from "../../src/paymaster"
 import { checkBalance, getBundlerUrl, getConfig, topUp } from "../utils"
-import { polygon } from "viem/chains"
-import { BiconomyFactoryAbi } from "../../src/account/abi/Factory"
 
 describe("Modules:Write", () => {
-  const nftAddress = "0x1758f42Af7026fBbB559Dc60EcE0De3ef81f665e"
   const {
     chain,
     chainId,
@@ -113,77 +107,6 @@ describe("Modules:Write", () => {
       expect(success).toBe("true")
     }
   }, 50000)
-
-  test("palgam", async () => {
-    const pr = "INSERT_PR_HERE"
-    const bundlerUrl =
-      "https://bundler.biconomy.io/api/v2/137/wde62ibhs.HI7fopYh-iJkl-12Io-af80-dsh90f74b78Cv"
-    const entryPoint = "0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789"
-
-    const account = privateKeyToAccount(pr)
-    const client = createWalletClient({
-      account,
-      chain: polygon,
-      transport: http()
-    })
-    const eoa = client.account.address
-
-    const smartAccount = await createSmartAccountClient({
-      signer: client,
-      bundlerUrl
-    })
-
-    const ecdsaModule = await createECDSAOwnershipValidationModule({
-      signer: client
-    })
-
-    const owner = await ecdsaModule.getAddress()
-    const smartAccountAddressFromSDK = await smartAccount.getAccountAddress()
-    const initCodeFromSDK = await smartAccount.getAccountInitCode()
-
-    console.log({ owner })
-
-    const publicClient = createPublicClient({
-      chain: polygon,
-      transport: http()
-    })
-
-    const factoryContract = getContract({
-      address: DEFAULT_BICONOMY_FACTORY_ADDRESS,
-      abi: BiconomyFactoryAbi,
-      client: { public: publicClient, wallet: client }
-    })
-
-    function getInitCode(walletAddress): Hex {
-      const encodedFunctionData = encodeFunctionData({
-        abi: BiconomyFactoryAbi,
-        functionName: "deployCounterFactualAccount",
-        args: [
-          owner,
-          `0x2ede3bc0000000000000000000000000${walletAddress.slice(2, 42)}`,
-          BigInt(0)
-        ]
-      })
-
-      const initCode =
-        DEFAULT_BICONOMY_FACTORY_ADDRESS + encodedFunctionData.substring(2)
-      return initCode as Hex
-    }
-
-    const initCodeFromFactory = getInitCode(eoa)
-
-    const smartAccountAddressFromContracts =
-      await factoryContract.read.getAddressForCounterFactualAccount([
-        owner,
-        getInitCode(eoa),
-        BigInt(0)
-      ])
-
-    expect(initCodeFromSDK.toUpperCase()).toBe(
-      initCodeFromFactory.toUpperCase()
-    )
-    expect(smartAccountAddressFromSDK).toBe(smartAccountAddressFromContracts)
-  })
 
   test.skip("should use MultichainValidationModule to mint an NFT on two chains with sponsorship", async () => {
     const nftAddress: Hex = "0x1758f42Af7026fBbB559Dc60EcE0De3ef81f665e"
